src/backend/requirements.txt
Flask>=3.0.0
Flask-Cors>=4.0.0
supabase>=2.4.0
python-dotenv>=1.0.1
requests>=2.31.0
pytest>=8.0.0
asgiref>=3.10.0